Separated analyzers using Collections into independent headers
